Conflicts:
	src/MapGenerator.cpp
	src/TemplateFactory.h

+++ src/MapGenerator.cpp
@@ -14,7 +14,7 @@ PrimitivaMountains::PrimitivaMountains(const DataBase &db, float power) : IMapGe
 #define GEN_OCT 5
 float PrimitivaMountains::flatness(float tx, float ty)
 {
-  return (PerlinNoise2D(tx, ty, 2, 2, GEN_OCT) + 1) / 2.f;
+  return (static_cast<float>(PerlinNoise2D(tx, ty, 2, 2, GEN_OCT)) + 1.0f) / 2.f;
 }
 float PrimitivaMountains::dens(float tx, float ty, float tz)
 {
@@ -29,7 +29,7 @@ float PrimitivaMountains::dens(float tx, float ty, float tz)
 }
 float cluster(float tx, float ty, float tz)
 {
-  return PerlinNoise3D(tx / 3.f, ty / 3.f, tz / 3.f, 1 + 5, 2, 1);
+  return static_cast<float>(PerlinNoise3D(tx / 3.f, ty / 3.f, tz / 3.f, 1 + 5, 2, 1));
 }
 // examples:
 // 0.2 = nothing
